Create a dom.js file
